+++ spec/forms/teams/players/create_form_spec.rb
@@ -6,6 +6,9 @@ describe Teams::Players::CreateForm, type: :service do
   let!(:instance) { described_class.new }
   let!(:seasons_team) { create :seasons_team }
   let!(:player) { create :player }
+  let!(:week) { create :week, season: seasons_team.season, status: 'active' }
+
+  before { create :game, week: week, home_season_team: seasons_team }
 
   context 'for invalid params' do
     let(:params) { { player_id: player.id, seasons_team_id: seasons_team.id, price_cents: 'abs' } }
@@ -21,8 +24,11 @@ describe Teams::Players::CreateForm, type: :service do
       { player_id: player.id, seasons_team_id: seasons_team.id, price_cents: 123, shirt_number: 1, form: 5.0 }
     }
 
-    it 'creates teams_player', :aggregate_failures do
-      expect { form }.to change(Teams::Player, :count).by(1)
+    it 'creates teams_player and games_player', :aggregate_failures do
+      expect { form }.to(
+        change(Teams::Player, :count).by(1)
+          .and(change(Games::Player, :count).by(1))
+      )
       expect(form[:errors]).to be_nil
     end
   end